requirements.txt
# Core ML and Data Science
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# FastAPI and Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6

# MLflow and Model Management
mlflow>=2.5.0
mlflow-pytorch>=2.5.0

# Monitoring and Metrics
prometheus-client>=0.17.0
structlog>=23.1.0

# Database and Caching
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=4.6.0

# Data Processing
pyarrow>=12.0.0
joblib>=1.3.0

# Configuration and Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development and Linting
black>=23.7.0
ruff>=0.0.284
mypy>=1.5.0
pre-commit>=3.3.0

# Jupyter for Development
jupyter>=1.0.0
jupyterlab>=4.0.0
ipykernel>=6.25.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Additional Utilities
requests>=2.31.0
aiofiles>=23.1.0
python-dateutil>=2.8.0
